+++ R/fuzzy_geom_set_operations.R
@@ -85,6 +85,25 @@ spa_flatten <- function(pcol) {
   new("pcollection", supp = pcol@supp, pgos = pgos)
 }
 
+#' Computes the standard union of a list of spatial plateau objects
+#' 
+#' @noRd
+internal_union_list <- function(pgos, type) {
+  if(length(pgos) == 0) {
+    create_empty_pgeometry(type)
+  } else if(length(pgos) == 1) {
+    pgos[[1]]
+  } else {
+    result <- pgos[[1]]
+    for(pgo in pgos[2:length(pgos)]){
+      result <- spa_union(result, pgo, utype = "max", as_pcomposition = FALSE)
+      if(fsr_is_empty(result)){
+        result <- create_empty_pgeometry(type)
+      }
+    }
+    result
+  }
+}
 
 #' @title Convert a spatial plateau collection into a spatial plateau composition
 #'
@@ -132,6 +151,7 @@ spa_flatten <- function(pcol) {
 #' @import sf
 #' @export
 pcollection_to_pcomposition <- function(pcol) {
+  # helper function that returns a quadruple of spatial plateau objects from a pcollection
   project <- function(pcol) {
     types <- sapply(pcol@pgos, spa_get_type)
     ppoint <- pcol@pgos[types == "PLATEAUPOINT"]
@@ -201,6 +221,202 @@ pcollection_to_pcomposition <- function(pcol) {
   result
 }
 
+#' Handles the result from crisp intersection and crisp difference operations in spatial plateau set operations
+#' 
+#' @noRd
+result_handler <- function(type1, type2, obj, md, comps, check_compatibility = FALSE) {
+  # helper function that adds a new component into a given list
+  append_comps <- function(obj, md, comps){
+    comp <- new("component", obj = obj, md = md)
+    append(comps, comp)
+  }
+  
+  if(!st_is_empty(obj) && md > 0 && md <= 1) {
+    geom_type <- st_geometry_type(obj)
+    # the compatibility is only checked during the intersection of components in the first phase of a geometric set operation (union and difference)
+    # when the types are equal and the resulting object is of same dimension as the spatial plateau objects, it means that we need to add the resulting object in the result
+    # otherwise this means that the resulting object is of lower dimension
+    # thus, we can don't add it in the list of components since this part of object will be present in a later point (during the second phase of the geometric set operation)
+    if(check_compatibility) {
+      if(type1 == type2 && type1 %in% c("PLATEAUPOINT", "PLATEAULINE", "PLATEAUREGION") && is_compatible(obj, type1)) {
+        if(type1 == "PLATEAUPOINT") {
+          comps$point <- append_comps(obj, md, comps$point)
+        } else if(type1 == "PLATEAULINE") {
+          comps$line <- append_comps(obj, md, comps$line)
+        } else if(type1 == "PLATEAUREGION") {
+          comps$region <- append_comps(obj, md, comps$region)
+        }
+      }
+    } else {
+      if(geom_type == "GEOMETRYCOLLECTION") {
+        comps$point <- append_comps(obj_union(obj, "POINT"), md, comps$point)
+        comps$line <- append_comps(obj_union(obj, "LINESTRING"), md, comps$line)
+        comps$region <- append_comps(obj_union(obj, "POLYGON"), md, comps$region)
+      } else if(geom_type %in% c("POINT", "MULTIPOINT")) {
+        comps$point <- append_comps(obj, md, comps$point)
+      } else if(geom_type %in% c("LINESTRING", "MULTILINESTRING")) {
+        comps$line <- append_comps(obj, md, comps$line)
+      } else if(geom_type %in% c("POLYGON", "MULTIPOLYGON")) {
+        comps$region <- append_comps(obj, md, comps$region)
+      }
+    }
+  }
+  comps
+}
+
+#' Builds a spatial plateau object corresponding to the result of a spatial plateau set operation between homogeneous data types
+#' 
+#' @noRd
+create_result_set_op <- function(comps, as_pcomposition, type) {
+  # we get the number of different spatial plateau data types required to compose the result
+  non_empty <- sum(sapply(list(comps$point, comps$line, comps$region), length) > 0)
+  if(non_empty >= 2) {
+    # in this case, only a plateau composition can be built
+    result <- create_empty_pgeometry("PLATEAUCOMPOSITION")
+    spa_add_internal(result, c(comps$point, comps$line, comps$region))
+  } else if(non_empty == 1 && !as_pcomposition) {
+    if(length(comps$point)) {
+      result <- create_empty_pgeometry("PLATEAUPOINT")
+      spa_add_internal(result, comps$point)
+    } else if(length(comps$line)) {
+      result <- create_empty_pgeometry("PLATEAULINE")
+      spa_add_internal(result, comps$line)
+    } else if(length(comps$region)) {
+      result <- create_empty_pgeometry("PLATEAUREGION")
+      spa_add_internal(result, comps$region)
+    }
+  } else if(non_empty == 1 && as_pcomposition) {
+    result <- create_empty_pgeometry("PLATEAUCOMPOSITION")
+    if(length(comps$point)) {
+      spa_add_internal(result, comps$point)
+    } else if(length(comps$line)) {
+      spa_add_internal(result, comps$line)
+    } else if(length(comps$region)) {
+      spa_add_internal(result, comps$region)
+    }
+  } else if(non_empty == 0 && as_pcomposition) {
+    create_empty_pgeometry("PLATEAUCOMPOSITION")
+  } else if(non_empty == 0 && !as_pcomposition) {
+    create_empty_pgeometry(type)
+  }
+}
+
+#' Computes spatial plateau set operation when at least of one operands is a plateau composition or plateau collection (i.e., a heterogeneous data type)
+#' 
+#' @noRd
+heterogeneous_geom_comp <- function(pgo1, pgo2, sigma, beta, as_pcomposition) {
+  # helper function that tries to simplify the structure of a spatial plateau object (e.g., if a composition has only one sub-object, and it will be simplified to such sub-object)
+  simplify <- function(pgo) {
+    if(fsr_is_empty(pgo)) {
+      return(pgo)
+    }
+    type <- spa_get_type(pgo)
+    if(type == "PLATEAUCOMPOSITION") {
+      triple <- c(pgo@ppoint, pgo@pline, pgo@pregion)
+      mask_empty_objs <- sapply(triple, fsr_is_empty)
+      n_empty_objs <- sum(mask_empty_objs)
+      if(n_empty_objs == 2) {
+        if(!mask_empty_objs[1]) {
+          pgo@ppoint
+        }else if(!mask_empty_objs[2]) {
+          pgo@pline
+        }else if(!mask_empty_objs[3]) {
+          pgo@pregion
+        }
+      } else {
+        pgo
+      }
+    }else if(type == "PLATEAUCOLLECTION") {
+      mask_empty_objs <- sapply(pgo@pgos, fsr_is_empty)
+      n_non_empty_objs <- length(pgo@pgos) - sum(mask_empty_objs)
+      if(n_non_empty_objs == 1) {
+        pgo@pgos[!mask_empty_objs][[1]]
+      } else {
+        pgo
+      }
+    } else {
+      pgo
+    }
+  }
+  
+  if(inherits(pgo1, "pcomposition") && inherits(pgo2, "pcomposition")) {
+    spgo1 <- simplify(pgo1)
+    spgo2 <- simplify(pgo2)
+    
+    if(inherits(spgo1, c("ppoint", "pline", "pregion")) && inherits(spgo2, c("ppoint", "pline", "pregion"))){
+      # if we successfully simplified BOTH objects, then we can simply call the underlying geometric set operation on such simplified objects
+      return(sigma(spgo1, spgo2, beta, as_pcomposition = as_pcomposition))
+    }
+    # otherwise, we have to compute the combination matrix
+    cm <- combination_matrix(pgo1, pgo2, sigma, beta, as_pcomposition = TRUE)
+    agg_combination_matrix(cm, as_pcomposition)
+  } else if(inherits(pgo1, c("ppoint", "pline", "pregion")) && inherits(pgo2, "pcomposition")) {
+    # we treat the other object as a pcomposition object and recursively call this function (it will enter in its first case)
+    pgo1 <- create_pgeometry(list(pgo1), "PLATEAUCOMPOSITION", is_valid = FALSE)
+    sigma(pgo1, pgo2, beta, as_pcomposition = as_pcomposition)
+  } else if(inherits(pgo1, "pcomposition") && inherits(pgo2, c("ppoint", "pline", "pregion"))) {
+    pgo2 <- create_pgeometry(list(pgo2), "PLATEAUCOMPOSITION", is_valid = FALSE)
+    sigma(pgo1, pgo2, beta, as_pcomposition = as_pcomposition)
+  } else if(inherits(pgo1, "pcollection") && inherits(pgo2, "pcollection")) {
+    spgo1 <- simplify(pgo1)
+    spgo2 <- simplify(pgo2)
+    
+    if(inherits(spgo1, c("ppoint", "pline", "pregion")) && inherits(spgo2, c("ppoint", "pline", "pregion"))) {
+      # if we successfully simplified BOTH objects, then we can simply call the underlying geometric set operation on such simplified objects
+      return(sigma(spgo1, spgo2, beta, as_pcomposition = as_pcomposition))
+    }
+    # otherwise, we have to transform those object as pcomposition objects and recursively call this function
+    pgo1 <- pcollection_to_pcomposition(pgo1)
+    pgo2 <- pcollection_to_pcomposition(pgo2)
+    sigma(pgo1, pgo2, beta, as_pcomposition = as_pcomposition)
+  } else if(inherits(pgo1, "pcollection") && inherits(pgo2, c("ppoint", "pline", "pregion", "pcomposition"))) {
+    pgo1 <- pcollection_to_pcomposition(pgo1)
+    sigma(pgo1, pgo2, beta, as_pcomposition = as_pcomposition)
+  } else if(inherits(pgo1, c("ppoint", "pline", "pregion", "pcomposition")) && inherits(pgo2, "pcollection")) {
+    pgo2 <- pcollection_to_pcomposition(pgo2)
+    sigma(pgo1, pgo2, beta, as_pcomposition = as_pcomposition)
+  }
+}
+
+#' Computes the combination matrix of two plateau composition objects
+#' 
+#' @noRd
+combination_matrix <- function(pgo1, pgo2, sigma, beta, as_pcomposition = TRUE) {
+  triple1 <- c(pgo1@ppoint, pgo1@pline, pgo1@pregion)
+  triple2 <- c(pgo2@ppoint, pgo2@pline, pgo2@pregion)
+  rep1 <- rep(triple1, each = 3)
+  mapply(sigma, rep1, triple2, beta, as_pcomposition)
+}
+
+#' Aggregates the combination matrix and returns a single spatial plateau object (possibly a plateau composition)
+#' 
+#' @noRd
+agg_combination_matrix <- function(cm, as_pcomposition = FALSE) {
+  ppoints <- lapply(cm, attr, "ppoint")
+  presult <- internal_union_list(ppoints[!sapply(ppoints, fsr_is_empty)], "PLATEAUPOINT")
+  
+  plines <- lapply(cm, attr, "pline")
+  lresult <- internal_union_list(plines[!sapply(plines, fsr_is_empty)], "PLATEAULINE")
+  
+  pregions <- lapply(cm, attr, "pregion")
+  rresult <- internal_union_list(pregions[!sapply(pregions, fsr_is_empty)], "PLATEAUREGION")
+  
+  mask_empty_objs <- sapply(c(presult, lresult, rresult), fsr_is_empty)
+  n_empty_objs <- sum(mask_empty_objs)
+  if(n_empty_objs == 2 && !as_pcomposition) {
+    non_empty_objs <- c(presult, lresult, rresult)[!mask_empty_objs][[1]]
+    non_empty_objs
+  } else {
+    supp <- st_union(st_sfc(presult@supp, lresult@supp, rresult@supp))[[1]]
+    agg_cm <- create_empty_pgeometry("PLATEAUCOMPOSITION")
+    agg_cm@supp <- supp
+    agg_cm@ppoint <- presult
+    agg_cm@pline <- lresult
+    agg_cm@pregion <- rresult
+    agg_cm
+  }
+}
+
 #' @title Fuzzy geometric set operations
 #'
 #' @description Fuzzy geometric set operations are given as a family of functions that implements spatial plateau set operations.
@@ -209,11 +425,12 @@ pcollection_to_pcomposition <- function(pcol) {
 #'
 #' @usage
 #'
-#' spa_intersection(pgo1, pgo2, itype = "min")
+#' spa_intersection(pgo1, pgo2, itype = "min", as_pcomposition = FALSE)
 #'
 #' @param pgo1 A `pgeometry` object of any type.
-#' @param pgo2 A `pgeometry` object of the same type of `pgo1`.
+#' @param pgo2 A `pgeometry` object of any type.
 #' @param itype A character value that indicates the name of a function implementing a t-norm. The default value is `"min"`, which is the standard operator of the intersection.
+#' @param as_pcomposition A logical value; if `TRUE`, it returns a spatial plateau composition object.
 #' 
 #' @name fsr_geometric_operations
 #'
@@ -253,6 +470,7 @@ pcollection_to_pcomposition <- function(pcol) {
 #' @references
 #'
 #' [Carniel, A. C.; Schneider, M. Spatial Plateau Algebra: An Executable Type System for Fuzzy Spatial Data Types. In Proceedings of the 2018 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2018), pp. 1-8, 2018.](https://ieeexplore.ieee.org/document/8491565)
+#' [Carniel, A. C.; Schneider, M. Spatial Data Types for Heterogeneously Structured Fuzzy Spatial Collections and Compositions. In Proceedings of the 2020 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2020), pp. 1-8, 2020.](https://ieeexplore.ieee.org/document/9177620)
 #'
 #' @examples
 #'
@@ -262,9 +480,9 @@ pcollection_to_pcomposition <- function(pcol) {
 #' pts2 <- rbind(c(1, 1), c(2, 3), c(2, 1))
 #' pts3 <- rbind(c(2, 2), c(3, 3))
 #'
-#' cp1 <- component_from_sfg(st_multipoint(pts1), 0.3)
-#' cp2 <- component_from_sfg(st_multipoint(pts2), 0.6)
-#' cp3 <- component_from_sfg(st_multipoint(pts3), 1)
+#' cp1 <- create_component(st_multipoint(pts1), 0.3)
+#' cp2 <- create_component(st_multipoint(pts2), 0.6)
+#' cp3 <- create_component(st_multipoint(pts3), 1)
 #' 
 #' pp1 <- create_pgeometry(list(cp1, cp2, cp3), "PLATEAUPOINT")
 #' 
@@ -272,9 +490,9 @@ pcollection_to_pcomposition <- function(pcol) {
 #' pts5 <- rbind(c(2, 3), c(1.2, 1.9), c(2, 1))
 #' pts6 <- rbind(c(3, 1), c(1.5, 0.5))
 #' 
-#' cp4 <- component_from_sfg(st_multipoint(pts4), 0.4)
-#' cp5 <- component_from_sfg(st_multipoint(pts5), 1)
-#' cp6 <- component_from_sfg(st_multipoint(pts6), 0.7)
+#' cp4 <- create_component(st_multipoint(pts4), 0.4)
+#' cp5 <- create_component(st_multipoint(pts5), 1)
+#' cp6 <- create_component(st_multipoint(pts6), 0.7)
 #' 
 #' pp2 <- create_pgeometry(list(cp4, cp5, cp6), "PLATEAUPOINT")
 #' 
@@ -288,37 +506,39 @@ pcollection_to_pcomposition <- function(pcol) {
 #'
 #' @import sf
 #' @export
-spa_intersection <- function(pgo1, pgo2, itype = "min"){
-
-  if(pgo1@type != pgo2@type){
-    stop("Different spatial plateau data types.", call. = FALSE)
-  }
-
-  sigma <- match.fun(itype)
-  result <- create_empty_pgeometry(pgo1@type)
-  lcomps <- vector("list")
-
-  for(comp1 in pgo1@component){
-    obj_comp_p1 <- comp1@obj
-    md_comp_p1 <- comp1@md
-
-    for(comp2 in pgo2@component){
-      obj_comp_p2 <- comp2@obj
-      md_comp_p2 <- comp2@md
-
-      result_md = sigma(md_comp_p1, md_comp_p2)
-
-      sf_result <- st_intersection(obj_comp_p1, obj_comp_p2)
-
-      # check geom and pgo
-      lcomps <- append_valid_comps(sf_result, result , result_md, lcomps)
+spa_intersection <- function(pgo1, pgo2, itype = "min", as_pcomposition = FALSE) {
+  # TODO add a short-circuit
+  
+  beta <- match.fun(itype)
+  comps <- list(point = list(), line = list(), region = list())
+  type1 <- spa_get_type(pgo1)
+  type2 <- spa_get_type(pgo2)
+  
+  if(type1 %in% c("PLATEAUPOINT", "PLATEAULINE", "PLATEAUREGION") && type2 %in% c("PLATEAUPOINT", "PLATEAULINE", "PLATEAUREGION")) {
+    # computing the intersection between homogeneous spatial plateau objects
+    for(comp1 in pgo1@component) {
+      for(comp2 in pgo2@component) {
+        result_md <- beta(comp1@md, comp2@md)
+        result_obj <- st_intersection(comp1@obj, comp2@obj)
+        comps <- result_handler(type1, type2, result_obj, result_md, comps, FALSE)
       }
-  }
-
-  if(length(lcomps) > 0){
-    spa_add_component(result, lcomps)
+    }
+    ftype <- final_data_type(type1, type2)
+    create_result_set_op(comps, as_pcomposition, ftype)
   } else {
-    result
+    sigma <- match.fun("spa_intersection")
+    heterogeneous_geom_comp(pgo1, pgo2, sigma, beta = itype, as_pcomposition = as_pcomposition)
+  }
+}
+
+#' @noRd
+is_pline_pregion_case <- function(type1, type2) {
+  if(type1 == "PLATEAULINE" && type2 == "PLATEAUREGION") {
+    TRUE
+  } else if(type1 == "PLATEAUREGION" && type2 == "PLATEAULINE") {
+    TRUE
+  } else {
+    FALSE
   }
 }
 
@@ -326,108 +546,122 @@ spa_intersection <- function(pgo1, pgo2, itype = "min"){
 #' 
 #' @usage
 #' 
-#' spa_union(pgo1, pgo2, utype = "max")
+#' spa_union(pgo1, pgo2, utype = "max", as_pcomposition = FALSE)
 #' 
 #' @param utype A character value that refers to a t-conorm. The default value is `"max"`, which is the standard operator of the union.
 #' 
 #' @import sf
 #' @export
-spa_union <- function(pgo1, pgo2, utype = "max"){
-
-  if(pgo1@type != pgo2@type){
-    stop("Different spatial plateau data types.", call. = FALSE)
-  }
-
+spa_union <- function(pgo1, pgo2, utype = "max", as_pcomposition = FALSE){
+  
   tau <- match.fun(utype)
-  result <- create_empty_pgeometry(pgo1@type)
-  lcomps <- vector("list")
-
-  supp_intersected <- st_intersection(pgo1@supp, pgo2@supp)
-
-  for(comp1 in pgo1@component){
-    obj_comp_p1 <- comp1@obj
-    md_comp_p1 <- comp1@md
-
+  comps <- list(point = list(), line = list(), region = list())
+  type1 <- spa_get_type(pgo1)
+  type2 <- spa_get_type(pgo2)
+  
+  if(type1 %in% c("PLATEAUPOINT", "PLATEAULINE", "PLATEAUREGION") && type2 %in% c("PLATEAUPOINT", "PLATEAULINE", "PLATEAUREGION")){
+    
+    if(!is_pline_pregion_case(type1, type2)){
+      supp_intersected <- st_intersection(pgo1@supp, pgo2@supp)
+    }
+    
+    for(comp1 in pgo1@component){
+      obj_comp_p1 <- comp1@obj
+      md_comp_p1 <- comp1@md
+      
+      for(comp2 in pgo2@component){
+        obj_comp_p2 <- comp2@obj
+        md_comp_p2 <- comp2@md
+        
+        result_md <- tau(md_comp_p1, md_comp_p2)
+        result_obj <- st_intersection(obj_comp_p1, obj_comp_p2)
+        comps <- result_handler(type1, type2, result_obj, result_md, comps, TRUE)
+      }
+      
+      # Special case between a spatial plateau region and a spatial plateau line
+      # TODO Conduct a performance evaluation to compare both strategies (probabily the first case is faster)
+      if(is_pline_pregion_case(type1, type2)){
+        sf_diff_1 <- st_difference(obj_comp_p1, pgo2@supp)
+      }else{
+        sf_diff_1 <- st_difference(obj_comp_p1, supp_intersected)
+      }
+      comps <- result_handler(type1, type2, sf_diff_1, md_comp_p1, comps, FALSE)
+    }
+    
     for(comp2 in pgo2@component){
       obj_comp_p2 <- comp2@obj
       md_comp_p2 <- comp2@md
-
-      result_md = tau(md_comp_p1, md_comp_p2)
-
-      sf_result <- st_intersection(obj_comp_p1, obj_comp_p2)
-
-      lcomps <- append_valid_comps(sf_result, result , result_md, lcomps)
+      
+      # Special case between a spatial plateau region and a spatial plateau line
+      if(is_pline_pregion_case(type1, type2)){
+        sf_diff_2 <- st_difference(obj_comp_p2, pgo1@supp)
+      }else{
+        sf_diff_2 <- st_difference(obj_comp_p2, supp_intersected)
+      }
+      comps <- result_handler(type1, type2, sf_diff_2, md_comp_p2, comps, FALSE)
     }
-
-    sf_diff_1 <- st_difference(obj_comp_p1, supp_intersected)
-    # check result type and appends to list if compatible
-    lcomps <- append_valid_comps(sf_diff_1, result, md_comp_p1, lcomps)
+    ftype <- final_data_type(type1, type2)
+    create_result_set_op(comps, as_pcomposition, ftype)
+  }else{
+    sigma <- match.fun("spa_union")
+    heterogeneous_geom_comp(pgo1, pgo2, sigma, beta = utype, as_pcomposition = as_pcomposition)
+    
   }
-
-  for(comp2 in pgo2@component){
-    obj_comp_p2 <- comp2@obj
-    md_comp_p2 <- comp2@md
-
-    sf_diff_3 <- st_difference(obj_comp_p2, supp_intersected)
-    # check result type and appends to list if compatible
-    lcomps <- append_valid_comps(sf_diff_3, result, md_comp_p2, lcomps)
-  }
-
-  if(length(lcomps) > 0){
-    spa_add_component(result, lcomps)
-  } else {
-    result
-  }
+  
 }
 
 #' @name fsr_geometric_operations
 #' 
 #' @usage
 #' 
-#' spa_difference(pgo1, pgo2, dtype = "f_diff")
+#' spa_difference(pgo1, pgo2, dtype = "f_diff", as_pcomposition = FALSE)
 #' 
 #' @param dtype A character value that indicates the name of a difference operator. The default value is `"f_diff"`, which implements the standard fuzzy difference.
 #' 
 #' @import sf
 #' @export
-spa_difference <- function(pgo1, pgo2, dtype = "f_diff"){
-
-  if(pgo1@type != pgo2@type){
-    stop("Different spatial plateau data types.", call. = FALSE)
-  }
-
+spa_difference <- function(pgo1, pgo2, dtype = "f_diff", as_pcomposition = FALSE){
+  
   nu <- match.fun(dtype)
-  result <- create_empty_pgeometry(pgo1@type)
-  lcomps <- vector("list")
-
-  supp_intersected <- st_intersection(pgo1@supp, pgo2@supp)
-
-  for(comp1 in pgo1@component){
-    obj_comp_p1 <- comp1@obj
-    md_comp_p1 <- comp1@md
-
-    for(comp2 in pgo2@component){
-      obj_comp_p2 <- comp2@obj
-      md_comp_p2 <- comp2@md
-
-      result_md = nu(md_comp_p1, md_comp_p2)
-
-      sf_result <- st_intersection(obj_comp_p1, obj_comp_p2)
-
-      lcomps <- append_valid_comps(sf_result, result , result_md, lcomps)
+  comps <- list(point = list(), line = list(), region = list())
+  type1 <- spa_get_type(pgo1)
+  type2 <- spa_get_type(pgo2)
+  
+  if(type1 %in% c("PLATEAUPOINT", "PLATEAULINE", "PLATEAUREGION") && type2 %in% c("PLATEAUPOINT", "PLATEAULINE", "PLATEAUREGION")){
+    
+    if(!is_pline_pregion_case(type1, type2)){
+      supp_intersected <- st_intersection(pgo1@supp, pgo2@supp)
     }
-
-    sf_diff_1 <- st_difference(obj_comp_p1, supp_intersected)
-
-    # check result type and appends to list if compatible
-    lcomps <- append_valid_comps(sf_diff_1, result, md_comp_p1, lcomps)
+    
+    for(comp1 in pgo1@component){
+      obj_comp_p1 <- comp1@obj
+      md_comp_p1 <- comp1@md
+      
+      for(comp2 in pgo2@component){
+        obj_comp_p2 <- comp2@obj
+        md_comp_p2 <- comp2@md
+        
+        result_md = nu(md_comp_p1, md_comp_p2)
+        result_obj = st_intersection(obj_comp_p1, obj_comp_p2)
+        comps <- result_handler(type1, type2, result_obj, result_md, comps, TRUE)
+        
+      }
+      # Special case between a spatial plateau region and a spatial plateau line
+      if(is_pline_pregion_case(type1, type2)){
+        sf_diff_1 <- st_difference(obj_comp_p1, pgo2@supp)
+      }else{
+        sf_diff_1 <- st_difference(obj_comp_p1, supp_intersected)
+      }
+      comps <- result_handler(type1, type2, sf_diff_1, md_comp_p1, comps, FALSE)
+    }
+    ftype <- final_data_type(type1, type2)
+    create_result_set_op(comps, as_pcomposition, ftype)
+    
+  }else{
+    sigma <- match.fun("spa_difference")
+    heterogeneous_geom_comp(pgo1, pgo2, sigma, beta = dtype, as_pcomposition = as_pcomposition)
   }
-
-  if(length(lcomps) > 0){
-    spa_add_component(result, lcomps)
-  } else {
-    result
-  }
+  
 }
 
 #' @name fsr_geometric_operations
@@ -443,46 +677,19 @@ spa_difference <- function(pgo1, pgo2, dtype = "f_diff"){
 #' @export
 spa_common_points <- function(pline1, pline2, itype = "min"){
   
-  if(pline1@type != pline2@type){
-    stop("Different Spatial Plateau Types.", call. = FALSE)
+  .Deprecated("spa_intersection")
+  
+  type1 <- spa_get_type(pline1)
+  type2 <- spa_get_type(pline2)
+  
+  if(type1 != type2){
+    stop("Different spatial plateau data types.", call. = FALSE)
   }
   
-  sigma <- match.fun(itype)
-  result <- create_empty_pgeometry(pline1@type)
-  lcomps <- vector("list")
+  result <- spa_intersection(pline1, pline2, itype = itype, as_pcomposition = TRUE)
   
-  for(comp1 in pline1@component){
-    obj_comp_p1 <- comp1@obj
-    md_comp_p1 <- comp1@md
-    
-    for(comp2 in pline2@component){
-      obj_comp_p2 <- comp2@obj
-      md_comp_p2 <- comp2@md
-      
-      result_md = sigma(md_comp_p1, md_comp_p2)
-      
-      sf_result <- st_intersection(obj_comp_p1, obj_comp_p2)
-      
-      if(!st_is_empty(sf_result) && st_geometry_type(sf_result) %in% c("POINT", "MULTIPOINT")){
-        result_comp <- new("component", obj = sf_result, md = result_md)
-        lcomps <- append(lcomps, result_comp)
-      } else if(!st_is_empty(sf_result) && st_geometry_type(sf_result) == "GEOMETRYCOLLECTION") {
-        type_geom = get_counter_ctype(pline1)
-        union_obj <- st_union(st_collection_extract(sf_result, type = type_geom))
-        if(inherits(union_obj, "sfc")) {
-          union_obj <- union_obj[[1]]
-        }
-        result_comp <- new("component", obj = union_obj, md = result_md)
-        lcomps <- append(lcomps, result_comp)
-      }
-    }
-  }
+  result@ppoint
   
-  if(length(lcomps) > 0){
-    spa_add_component(result, lcomps)
-  } else {
-    result
-  }
 }
 
 #' @title Fuzzy difference operators